% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documented_objects.R
\name{BreakPoint}
\alias{BreakPoint}
\title{BreakPoint object}
\value{
\item{fragments}{A \code{\link[GenomicRanges]{GRanges}} object with read fragments.}
\item{deltas}{A \code{\link[GenomicRanges]{GRanges}} object with deltaWs.}
\item{breaks}{A \code{\link[GenomicRanges]{GRanges}} object containing the breakpoint coordinates.}
\item{counts}{A \code{\link[GenomicRanges]{GRanges}} object with the regions between breakpoints.}
\item{params}{A vector with parameters that were used to obtain the results.}
}
\description{
The \code{BreakPoint} object is output of the function \code{\link{runBreakpointr}} and is basically a list with various entries. The class() attribute of this list was set to "BreakPoint". Entries can be accessed with the list operators '[[]]' and '$'.
}
\seealso{
runBreakpointr
}

