% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakSeekr.R
\name{breakSeekr}
\alias{breakSeekr}
\title{Find breakpoints from deltaWs}
\usage{
breakSeekr(deltaWs, trim = 10, peakTh = 0.33, zlim = 3.291)
}
\arguments{
\item{deltaWs}{A \code{\link[GenomicRanges]{GRanges}} object with metadata column "deltaW" generated by \code{\link{deltaWCalculator}}.}

\item{trim}{Trim the upper and lower deltaWs in the dataset (e.g. 10 calculates SD of all values btwn 10th and 90th percentile).}

\item{peakTh}{Determines the fraction of the upper peak to use as the threshold (e.g. 0.33 is 1/3 of max(delta)).}

\item{zlim}{The degree that the value must be above the threshold to be included in the peaklist.}
}
\value{
A \code{\link[GenomicRanges]{GRanges}} object containing breakpoint coordinates with various metadata columns.
}
\description{
Find breakpoints from deltaWs by <TODO:DAVID here a description of how breakpoints are found>.
}
\author{
Ashley Sanders, David Porubsky, Aaron Taudt
}

