% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deltaWCalculator.R
\name{deltaWCalculator}
\alias{deltaWCalculator}
\title{Calculate deltaWs}
\usage{
deltaWCalculator(frags, reads.per.window = 100)
}
\arguments{
\item{frags}{A \code{\link[GenomicRanges]{GRanges}} with read fragments (see \code{\link{readBamFileAsGRanges}}).}

\item{reads.per.window}{Number of reads in each dynamic window.}
}
\value{
The input \code{frags} with additional meta-data columns.
}
\description{
This function will calculate deltaWs from a \code{\link[GenomicRanges]{GRanges}} object with read fragments.
}
\examples{
filepath <- system.file("extdata", "breakpointer", package="strandseqExampleData")
file <- list.files(filepath, full.names=TRUE)[1] 
fragments <- readBamFileAsGRanges(file, pairedEndReads=FALSE)
dw <- deltaWCalculator(fragments)

}
\author{
Aaron Taudt
}
\seealso{
readBamFileAsGRanges
}

