% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deltaWCalculator.R
\name{deltaWCalculatorVariousWindows}
\alias{deltaWCalculatorVariousWindows}
\title{Calculate deltaWs using various window sizes}
\usage{
deltaWCalculatorVariousWindows(frags, reads.per.window = 100,
  multi.sizes = c(2, 4, 6))
}
\arguments{
\item{frags}{A \code{\link{GRanges-class}} with read fragments (see \code{\link{readBamFileAsGRanges}}).}

\item{reads.per.window}{Number of reads in each dynamic window.}

\item{multi.sizes}{User defined multiplications of the original window size.}
}
\value{
The input \code{frags} with additional meta-data columns.
}
\description{
This function will calculate deltaWs from a \code{\link{GRanges-class}} object with read fragments.
}
\seealso{
deltaWCalculator
}
\author{
David Porubsky
}
