% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documented_objects.R
\docType{data}
\name{example_BreakPoints}
\alias{example_BreakPoints}
\title{BreakPoint objects for demonstration purposes}
\format{A list with \code{\link{BreakPoint}} objects.}
\description{
A list of \code{\link{BreakPoint}} objects for demonstration purposes in examples of package \pkg{\link{breakpointR}}.
}
\examples{
data(example_BreakPoints)
length(example_BreakPoints)
## Have a look at the first object
example_BreakPoints[[1]]
}
