% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{insertchr}
\alias{insertchr}
\title{Insert chromosome for in case it's missing}
\usage{
insertchr(gr)
}
\arguments{
\item{gr}{A \code{\link{GRanges-class}} object.}
}
\value{
The input \code{\link{GRanges-class}} object with an additional metadata column containing chromosome name with 'chr'.
}
\description{
Add two columns with transformed genomic coordinates to the \code{\link{GRanges-class}} object. This is useful for making genomewide plots.
}
