% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotBreakpointCoverage}
\alias{plotBreakpointCoverage}
\title{Plot breakpoint summary \pkg{\link{breakpointR}}}
\usage{
plotBreakpointCoverage(gr, MbPerRow = NULL)
}
\arguments{
\item{gr}{A \code{\link{GRanges-class}} object that contains column 'cov' - breakpoint coverage per genomic regions.}

\item{MbPerRow}{...}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot}} object.
}
\description{
This function takes breakpoint coverage per genomic ranges and plots them as peak along the genome.
}
\author{
David Porubsky
}
