% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotBreakpoints}
\alias{plotBreakpoints}
\title{Plotting functions for \pkg{\link{breakpointR}}}
\usage{
plotBreakpoints(files2plot, file = NULL)
}
\arguments{
\item{file}{Name of the file to plot to.}

\item{breakpoints}{A list of \code{\link{BreakPoint}} objects or a vector with files that contain such objects.}
}
\value{
A list with \code{\link[ggplot2:ggplot]{ggplot}} objects.
}
\description{
This function will produce several plots from a \code{\link{BreakPoint}} object.
}
\examples{
## Get an example file
exampleFolder <- system.file("extdata", "breakpointer", package="strandseqExampleData")
exampleFile <- list.files(exampleFolder, full.names=TRUE)[1]
## Run breakpointR
brkpts <- runBreakpointr(exampleFile, pairedEndReads=FALSE,
                         chromosomes=paste0('chr', c(1:22,'X')))
## Plot them
plotBreakpoints(brkpts)
}
\author{
David Porubsky, Aaron Taudt, Ashley Sanders
}

