% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotHeatmap}
\alias{plotHeatmap}
\title{Genome wide heatmap of template inheritance states}
\usage{
plotHeatmap(datapath, plotLibraries = NULL, file = NULL)
}
\arguments{
\item{datapath}{location of RData files storing reads and breakpoint coordinates}

\item{plotLibraries}{subset of libaries to be plotted, otherwise all libraries will be plotted}

\item{file}{name of the file to store libraries in}
}
\description{
Plot a genome wide heatmap of template inheritance states.
}
\author{
David Porubsky, Aaron Taudt, Ashley Sanders
}

