% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/writeBedFile.R
\name{writeBedFile}
\alias{writeBedFile}
\title{Generates a bedfile from inFile}
\usage{
writeBedFile(index, outputDirectory, fragments = NULL, deltaWs = NULL,
  breakTrack = NULL, maskedRegions = NULL, hotspots = NULL,
  breaksGraph = NULL, bin = NA, col = "103,139,139 243,165,97")
}
\arguments{
\item{index}{is used to index the bedfile(s)}

\item{outputDirectory}{is location to write bedfile(s)}

\item{fragments}{A \code{\link{GRanges}} object with strand and mapq metadata, such as that generated by \code{\link{bam2GRanges}}}

\item{deltaWs}{A \code{\link{GRanges}} object with metadata column "deltaW" generated by \code{\link{deltaWCalculator}}}

\item{breakTrack}{A \code{\link{GRanges}} object with metadata "genoT" (e.g. newBreaks) will write a bedtrack with refined breakpoints}

\item{maskedRegions}{A \code{\link{GRanges}} object}

\item{hotspots}{A \code{\link{GRanges}} object}

\item{breaksGraph}{A \code{\link{GRanges}} object}
}
\description{
Write a bedfile from Breakpoint.R files for upload on to UCSC Genome browser
}
\author{
Ashley Sanders, David Porubsky, Aaron Taudt
}

