% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakpointer.R
\name{breakpointer}
\alias{breakpointer}
\title{Main function for the \pkg{\link{breakpointR}} package}
\usage{
breakpointer(inputfolder, outputfolder, configfile = NULL, numCPU = 1,
  reuse.existing.files = FALSE, windowsize = 1e+06, binMethod = "size",
  pairedEndReads = FALSE, pair2frgm = FALSE, chromosomes = NULL,
  min.mapq = 10, trim = 10, peakTh = 0.33, zlim = 3.291,
  background = 0.02, minReads = 10, createCompositeFile = FALSE,
  WC.cutoff = 0.9, maskRegions = NULL, callHotSpots = FALSE)
}
\arguments{
\item{inputfolder}{Folder with BAM files.}

\item{outputfolder}{Folder to output the results. If it does not exist it will be created.}

\item{configfile}{A file specifying the parameters of this function (without \code{inputfolder}, \code{outputfolder} and \code{configfile}). Having the parameters in a file can be handy if many samples with the same parameter settings are to be run. If a \code{configfile} is specified, it will take priority over the command line parameters.}

\item{numCPU}{The numbers of CPUs that are used. Should not be more than available on your machine.}

\item{reuse.existing.files}{A logical indicating whether or not existing files in \code{outputfolder} should be reused.}

\item{windowsize}{The window size used to calculate deltaWs, either number of reads or genomic size depending on binMethod}

\item{binMethod}{Method used to calculate optimal number of reads in the window ("size", "reads"). Default = "size"}

\item{pairedEndReads}{Set to \code{TRUE} if you have paired-end reads in your file.}

\item{pair2frgm}{Set to \code{TRUE} if every paired-end read should be merged into a single fragment}

\item{chromosomes}{If only a subset of the chromosomes should be binned, specify them here.}

\item{min.mapq}{Minimum mapping quality when importing from BAM files.}

\item{background}{The amount of background introduced into the genotype test.}

\item{createCompositeFile}{Merge single cell data into a single file to increase breakpoint mapping resolution (do not use for SCE mapping)}

\item{WC.cutoff}{Percentage of WW or CC reads to consider chromosome being WW or CC}

\item{maskRegions}{List of regions to be excluded from the analysis (format: chromosomes start end)}

\item{callHotSpots}{Search for regions of high abundance of breakpoints in single cells (set createCompositeFile=FALSE)}
}
\value{
\code{NULL}
}
\description{
This function is an easy-to-use wrapper to find breakpoints with \code{\link[breakpointR]{runBreakpointr}} in parallel, write the results to file and find hotspots.
}
\examples{
## Get some example files
inputfolder <- system.file("extdata", "breakpointer", package="strandseqExampleData")
outputfolder <- file.path(tempdir(), "breakpointer_example")
## Run breakpointer
breakpointer(inputfolder, outputfolder, chromosomes=paste0("chr", c(1:22, "X")),
            pairedEndReads=FALSE)

}
\author{
Aaron Taudt, David Porubsky, Ashley Sanders
}

