% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakpointR-package.R
\docType{package}
\name{breakpointR-package}
\alias{breakpointR-package}
\alias{breakpointR}
\title{Breakpoint detection in Strand-Seq data}
\description{
This package implements functions for finding breakpoints, plotting and export of Strand-seq data.
}
\details{
The main function of this package is \code{\link{breakpointr}} and produces several plots and browser files. If you want to have more fine-grained control over the different steps check the vignette \href{../doc/breakpointR.pdf}{How to use breakpointR}.
}
\author{
David Porubsky, Ashley Sanders, Aaron Taudt
}
