% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidenceInterval.R
\name{confidenceInterval}
\alias{confidenceInterval}
\title{Estimate confidence intervals for breakpoints}
\usage{
confidenceInterval(breaks, fragments, background, conf = 0.99)
}
\arguments{
\item{breaks}{Genotyped breakpoints as outputted from function \code{\link{GenotypeBreaks}}.}

\item{fragments}{Read fragments from function \code{\link{readBamFileAsGRanges}}.}

\item{background}{The percent (e.g. 0.02 = 2\%) of background reads allowed for WW or CC genotype calls.}

\item{conf}{Desired confidence interval.}
}
\description{
Estimate confidence intervals for breakpoints by going outwards from the breakpoint read by read, and multiplying the probability that the read doesn't belong to the assigned segment.
}
\examples{
filepath <- system.file("extdata", "breakpointer", package="strandseqExampleData")
file <- list.files(filepath, full.names=TRUE)[1]
fragments <- readBamFileAsGRanges(file, pairedEndReads=FALSE)
dw <- deltaWCalculator(fragments)
breaks <- breakSeekr(dw)
gbreaks <- GenotypeBreaks(breaks, fragments)
confint <- confidenceInterval(gbreaks, fragments, background=0.02)

}

