% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createCompositeFile.R
\name{createCompositeFile}
\alias{createCompositeFile}
\title{Create composite Strand-seq file}
\usage{
createCompositeFile(file.list, chromosomes = NULL, pairedEndReads = TRUE,
  pair2frgm = FALSE, min.mapq = 10, remove.duplicate.reads = TRUE,
  WC.cutoff = 0.9, bg = 0.02)
}
\arguments{
\item{file.list}{TODO:DAVID}

\item{chromosomes}{If only a subset of the chromosomes should be binned, specify them here.}

\item{pairedEndReads}{Set to \code{TRUE} if you have paired-end reads in your file.}

\item{pair2frgm}{Set to \code{TRUE} if every paired-end read should be merged into a single fragment}

\item{min.mapq}{Minimum mapping quality when importing from BAM files.}

\item{remove.duplicate.reads}{A logical indicating whether or not duplicate reads should be kept.}

\item{WC.cutoff}{Percentage of WW or CC reads to consider chromosome being WW or CC}

\item{bg}{The amount of background introduced into the genotype test.}
}
\value{
TODO:DAVID
}
\description{
This function will move through BAM files in a folder, read in each individual file and go through each chromosome, 
determine if the chromosome is WW or CC based on WCcutoff, reverse complement all reads in the WW file,
append to a new composite file for that chromosome, order the composite file of each chromosome based on position.
}
\examples{
## Get some example files
inputfolder <- system.file("extdata", "breakpointer", package="strandseqExampleData")
files <- list.files(inputfolder, full.names=TRUE, pattern="bam$")
## Create the composite file
composite <- createCompositeFile(files, chromosomes=paste0('chr', c(1:22, 'X')),
                   pairedEndReads=FALSE)
                   
}
\author{
Ashley Sanders, David Porubsky
}

