% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/createCompositeFile.R
\name{createCompositeFile}
\alias{createCompositeFile}
\title{Function to create composite Strand-seq file}
\usage{
createCompositeFile(file.list, chromosomes = NULL, pairedEndReads = TRUE,
  min.mapq = 10, keep.duplicate.reads = FALSE, WC.cutoff = 0.9)
}
\arguments{
\item{chromosomes}{If only a subset of the chromosomes should be binned, specify them here.}

\item{pairedEndReads}{Set to \code{TRUE} if you have paired-end reads in your file.}

\item{min.mapq}{Minimum mapping quality when importing from BAM files.}

\item{keep.duplicate.reads}{A logical indicating whether or not duplicate reads should be kept.}

\item{WC.cutoff}{Percentage of WW or CC reads to consider chromosome being WW or CC}

\item{datapath}{Location where bam files to be processed are stored}
}
\description{
This script will move through .bam files in a folder
read in each individual file and go thru each chr
determine the chr is WW or CC based on WCcutoff
reverse compliment all reads in the WW file
append to a new composite file for that chr
order the composite file of each chr based on pos
}
\author{
Ashley Sanders, David Porubsky
}

