% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenotypeBreaks.R
\name{genotype.fisher}
\alias{genotype.fisher}
\title{Genotype read numbers}
\usage{
genotype.fisher(cReads, wReads, roiReads, background = 0.02,
  minReads = 10)
}
\arguments{
\item{cReads}{Number of Crick reads.}

\item{wReads}{Number of Watson reads.}

\item{roiReads}{Total number of reads.}

\item{background}{Parameter for frequency of background reads.}

\item{minReads}{Minimal number of reads to perform the test.}
}
\value{
A list with the $bestFit and $pval.
}
\description{
Genotype read numbers with Fisher's exact test.
}
\author{
Ashley Sanders, David Porubsky, Aaron Taudt
}
