% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenotypeBreaks.R
\name{GenotypeBreaks}
\alias{GenotypeBreaks}
\title{Genotype the space between breakpoints}
\usage{
GenotypeBreaks(breaks, fragments, background = 0.05, minReads = 10)
}
\arguments{
\item{breaks}{A \code{\link{GRanges-class}} object with breakpoint coordinates.}

\item{fragments}{A \code{\link{GRanges-class}} object with read fragments.}

\item{background}{The percent (e.g. 0.05 = 5\%) of background reads allowed for WW or CC genotype calls.}

\item{minReads}{The minimal number of reads between two breaks required for genotyping.}
}
\value{
A \code{\link{GRanges-class}} object with genotyped breakpoint coordinates.
}
\description{
Genotype the space between breakpoints with Fisher's exact test.
}
\examples{
## Get an example file 
exampleFolder <- system.file("extdata", "example_bams", package="breakpointRdata")
exampleFile <- list.files(exampleFolder, full.names=TRUE)[1]
## Load the file 
fragments <- readBamFileAsGRanges(exampleFile, pairedEndReads=FALSE)
## Calculate deltaW values
dw <- deltaWCalculator(fragments)
## Get significant peaks in deltaW values
breaks <- breakSeekr(dw)
## Genotype regions between breakpoints
gbreaks <- GenotypeBreaks(breaks, fragments)

}
\author{
David Porubsky, Ashley Sanders, Aaron Taudt
}
