% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GenotypeBreaks.R
\name{GenotypeBreaks}
\alias{GenotypeBreaks}
\title{Genotype the space between breakpoints}
\usage{
GenotypeBreaks(breaks, fragments, backG = 0.02, minReads = 10)
}
\arguments{
\item{breaks}{A \code{\link{GRanges}} object with breakpoint coordinates.}

\item{fragments}{A \code{\link{GRanges}} object with read fragments.}

\item{backG}{The percent (e.g. 0.02 = 2\%) of background reads allowed for WW or CC genotype calls.}

\item{minReads}{The minimal number of reads between two breaks required for genotyping.}
}
\description{
Genotype the space between breakpoints with Fisher's exact test
}
\author{
Ashley Sanders, David Porubsky, Aaron Taudt
}

