% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotBreakpoints}
\alias{plotBreakpoints}
\title{Plotting genome-wide ideograms \pkg{\link{breakpointR}}}
\usage{
plotBreakpoints(files2plot, file = NULL)
}
\arguments{
\item{files2plot}{A list of files that contains \code{\link{BreakPoint}} objects or a single \code{\link{BreakPoint}} object.}

\item{file}{Name of the file to plot to.}
}
\value{
A list with \code{\link[ggplot2:ggplot]{ggplot}} objects.
}
\description{
This function will create genome-wide ideograms from a \code{\link{BreakPoint}} object.
}
\examples{
## Get an example file
exampleFolder <- system.file("extdata", "example_results", package="strandseqExampleData")
exampleFile <- list.files(exampleFolder, full.names=TRUE)[1]
## Plot the file
plotBreakpoints(exampleFile)
}
\author{
David Porubsky, Aaron Taudt, Ashley Sanders
}
