% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotBreakpoints}
\alias{plotBreakpoints}
\title{Plotting function for BreakPointR
This function will take stored RData files and plot them into a single PDF file}
\usage{
plotBreakpoints(datapath, plotLibraries = NULL, file = NULL)
}
\arguments{
\item{datapath}{location of RData files storing reads and breakpoint coordinates}

\item{plotLibraries}{subset of libaries to be plotted, otherwise all libraries will be plotted}

\item{file}{name of the file to store libraries in}
}
\description{
Plotting function for BreakPointR
This function will take stored RData files and plot them into a single PDF file
}
\author{
David Porubsky, Aaron Taudt, Ashley Sanders
}

