% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotHeatmap}
\alias{plotHeatmap}
\title{Genome wide heatmap of template inheritance states}
\usage{
plotHeatmap(breakpoints, file = NULL)
}
\arguments{
\item{breakpoints}{A list of \code{\link{BreakPoint}} objects or a vector with files that contain such objects.}

\item{file}{Name of the file to plot to.}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot}} object of \code{NULL}, depending on option \code{file}.
}
\description{
Plot a genome wide heatmap of template inheritance states.
}
\examples{
## Get example BreakPoint objects to plot
data(example_BreakPoints)
## Plot the heatmap
plotHeatmap(example_BreakPoints)

}
\author{
David Porubsky, Aaron Taudt, Ashley Sanders
}

