% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/runBreakpointr.R
\name{runBreakpointr}
\alias{runBreakpointr}
\title{Wrapper function for breakpointR}
\usage{
runBreakpointr(input.data, pairedEndReads = TRUE, chromosomes = NULL,
  windowsize = 1e+06, scaleWindowSize = T, trim = 10, peakTh = 0.33,
  zlim = 3.291, bg = 0.02, min.mapq = 10, pair2frgm = FALSE,
  minReads = 20, maskRegions = NULL)
}
\arguments{
\item{input.data}{A file with aligned reads in BAM format.}

\item{pairedEndReads}{Set to \code{TRUE} if you have paired-end reads in your file.}

\item{chromosomes}{If only a subset of the chromosomes should be processed, specify them here.}

\item{windowsize}{The window size used to calculate deltaWs, either an integer or 'scale'.}

\item{trim}{The amount of outliers in deltaWs removed to calculate the stdev (10 will remove top 10\% and bottom 10\% of deltaWs).}

\item{peakTh}{The treshold that the peak deltaWs must pass to be considered a breakpoint.}

\item{zlim}{The number of stdev that the deltaW must pass the peakTh (ensures only significantly higher peaks are considered).}

\item{bg}{The amount of background introduced into the genotype test.}

\item{minReads}{The minimum number of reads required for genotyping.}

\item{maskRegions}{List of regions to be excluded from the analysis (format: chromosomes start end)}
}
\description{
This script will move through .bam files in a folder and perform several steps (see Details).
}
\details{
1. calculate deltaWs chromosome-by-chromsome
2. localize breaks that pass zlim above the threshold
3. genotype both sides of breaks to confirm whether strand state changes
4. write a file of _reads, _deltaWs and _breaks in a chr fold -> can upload on to UCSC Genome browser
5. write a file for each index with all chromosomes included -> can upload on to UCSC Genome browser
}
\author{
Ashley Sanders, David Porubsky, Aaron Taudt
}

