% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeBedFile.R
\name{writeBedFile}
\alias{writeBedFile}
\title{Generates a bedfile from inFile}
\usage{
writeBedFile(index, outputDirectory, fragments = NULL, deltaWs = NULL,
  breakTrack = NULL, confidenceIntervals = NULL, maskedRegions = NULL,
  hotspots = NULL, breaksGraph = NULL, col = "103,139,139 243,165,97")
}
\arguments{
\item{index}{A character used to name the bedfile(s).}

\item{outputDirectory}{Location to write bedfile(s).}

\item{fragments}{A \code{\link[GenomicRanges]{GRanges}} object with strand and mapq metadata, such as that generated by \code{\link{readBamFileAsGRanges}}}

\item{deltaWs}{A \code{\link[GenomicRanges]{GRanges}} object with metadata column "deltaW" generated by \code{\link{deltaWCalculator}}.}

\item{breakTrack}{A \code{\link[GenomicRanges]{GRanges}} object with metadata "genoT" (e.g. newBreaks) will write a bedtrack with refined breakpoints.}

\item{confidenceIntervals}{A \code{\link[GenomicRanges]{GRanges}} object with metadata "genoT" the same length as \code{breakTrack} (e.g. confint) will write a bedtrack with breakpoints confidence intervals.}

\item{maskedRegions}{A \code{\link[GenomicRanges]{GRanges}} object.}

\item{hotspots}{A \code{\link[GenomicRanges]{GRanges}} object.}

\item{breaksGraph}{A \code{\link[GenomicRanges]{GRanges}} object.}

\item{col}{A string with two RGB values separated by a space specifying the color for the two strands.}
}
\value{
\code{NULL}
}
\description{
Write a bedfile from Breakpoint.R files for upload on to UCSC Genome browser
}
\examples{
## Get an example file
exampleFolder <- system.file("extdata", "example_bams", package="strandseqExampleData")
exampleFile <- list.files(exampleFolder, full.names=TRUE)[1]
## Run breakpointR
brkpts <- runBreakpointr(exampleFile, pairedEndReads=FALSE, chromosomes='chr1')
## Write results to BED files
writeBedFile(index='testfile', outputDirectory=tempdir(), fragments=brkpts$fragments,
           deltaWs=brkpts$deltas, breakTrack=brkpts$breaks,
           confidenceIntervals=brkpts$confint)
           
}
\author{
Ashley Sanders, David Porubsky, Aaron Taudt
}
